home api slice optimized
